+++ pkg/deploy/bindata.go
@@ -96,7 +96,7 @@ func (fi bindataFileInfo) Sys() interface{} {
 	return nil
 }
 
-var _clusterDevelopmentPredeployJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x58\x5d\x8f\xa3\x36\x14\x7d\x9f\x5f\x81\x68\x25\x12\x29\x7c\x24\xda\x56\x9d\x79\x1b\xb5\xdd\x6a\xb4\xea\x76\xd4\x6c\xa7\x0f\x51\x1e\x1c\x73\x93\xf1\xc6\xd8\xc8\x36\xd9\xb2\x55\xfe\x7b\x65\x08\x24\x80\x21\x2c\x93\xb4\xa3\x55\xb3\xd2\x4a\xc0\xf5\xf5\xfd\x3a\xc7\xc7\xf3\xf7\x8d\x65\x59\x96\xfd\xad\xc4\xcf\x10\x21\xfb\xce\xb2\x9f\x95\x8a\xe5\x9d\xef\xe7\x6f\xbc\x08\x31\xb4\x81\x08\x98\xf2\xd0\xe7\x44\x80\x87\x79\x74\xf8\x26\xfd\x59\x30\xfd\xce\x0d\xa6\x6e\x30\xf5\x43\x88\x29\x4f\xb5\xdd\x07\x88\x62\x8a\x14\x78\x1f\x25\x67\xdf\xd8\x93\x7c\x07\xcc\x99\x02\xa6\x9e\x40\x48\xc2\x99\xde\x68\xea\x05\xfa\x5f\x61\x10\x23\x81\x22\x50\x20\xa4\x7d\x67\xe5\x61\xe5\x0b\x49\xe5\x39\x7b\xa7\xd2\x18\xb4\x8b\x15\xe7\xf4\xb0\xbe\xfc\x16\xc2\x1a\x25\x54\x3d\x21\x9a\x68\x9b\x35\xa2\x12\x4a\x8b\xfd\xe4\xc4\x31\x4d\xa4\x02\xf1\x1e\x45\xd0\xbe\x83\x54\x82\xb0\x8d\xdd\xe5\x60\x0e\x62\x47\x30\x3c\x0a\xc2\x30\x89\x11\x7d\x08\x87\xb9\x0b\x89\xdc\xfe\xcc\xb0\x48\x63\x45\x38\x9b\x83\x1a\x1e\xd9\x3a\xbe\x54\x50\xdb\xdd\xf0\x28\x22\xa4\xcb\x73\x1f\x86\x02\xa4\x7c\x14\xb0\x26\x7f\x0d\x73\x24\x78\xa2\x40\xb6\xaf\x45\x42\xa0\xf4\xcc\x1c\x2c\x96\x46\xd7\x3b\x06\xea\x02\x11\x7e\xe2\x62\x3b\x3c\xd5\x9b\x13\x77\xb6\x00\xc9\x13\x81\xb3\x7c\x17\xa5\x4d\xcd\x55\x2c\x78\x0c\x42\x11\x43\x55\xb2\xef\x28\x8f\x64\x1e\x23\xdc\x6c\x5e\xdd\x2a\x8f\xb7\xb6\x61\xc3\x78\x71\x04\xe8\xc8\x69\x94\xcd\x19\x2f\x6d\xe3\xda\x65\xe3\xed\xbe\xf2\x66\x5f\x6b\x1b\xcb\xc7\xcd\x0e\x61\xe7\xea\x5d\xea\x6d\x2d\x8a\xf8\x2b\xc1\x82\x4b\xbe\x56\xde\x7b\x50\xba\xfc\xfe\x8e\x08\x95\x20\x7a\x78\x94\xf5\x85\x94\x63\xa4\x0e\xd4\xb3\x28\x8a\xfc\x8b\xe0\x49\x3c\x1a\x7b\xc5\xc7\x65\x7d\x15\xe6\x2c\x24\xe5\xb2\xd3\x12\x60\xa2\x73\xae\x99\xa3\x98\x9c\x30\xdc\x2c\x98\x05\x6e\xf0\x83\x1b\x4c\x8d\x53\xf3\x85\x3d\x2d\x51\x50\x8d\x23\x7f\xdd\xa8\x7f\x5b\x5d\x17\x98\x33\x8c\xd4\xa8\x92\xca\x91\x09\x9d\xf1\xc4\x72\x5c\xa1\x0c\xb9\xb5\x57\x3e\x0b\xe1\x03\x5a\x51\xb8\x50\xd5\xaf\x59\x46\x54\x03\x69\xb5\x9a\x06\xce\x6a\x96\xe2\xd8\x8e\x2c\xe9\x76\x80\x91\xb0\x92\x37\x09\x47\x4e\x67\xf5\x9c\x89\xd5\xb7\x3d\xc6\xa0\xf2\x3e\xa1\x8d\x4e\x9d\x25\x94\x0e\xc4\x5e\x31\x23\x4e\x01\x42\xdf\x99\x58\x9d\x11\xe5\x65\xfb\xa2\xa1\xa9\xc1\xd5\x97\xc9\x8a\x81\x6a\x0c\x50\xeb\x28\x34\xe8\x3e\x06\x16\xca\xdf\x98\x91\xc5\xce\xf5\xa0\x16\x8c\x33\xb1\xca\xdc\x5b\xda\x7f\xad\xa6\x56\x76\x5a\x0e\x43\xd3\xd5\x51\x62\x38\xee\xfe\x47\xc9\x79\x94\xe4\x65\xfb\x6a\x50\x52\x04\x53\x41\x4b\xbf\xde\x14\x7c\xf1\x55\x62\xab\x1c\x90\xd3\x98\x72\x21\xdd\xab\xf7\xef\x20\x7d\xd2\xd2\xd5\xdf\xe9\xff\x2f\x74\xa4\x9e\x03\xbc\x02\x86\x98\x7a\xc8\xa1\x28\x93\x95\xc4\x82\x64\x17\x91\xd1\xd8\x2b\xbe\x19\x9b\x25\xb7\x49\x3b\xb4\xd7\x28\x22\x34\xd5\x3e\xef\xdb\x80\x58\x54\x4b\x2a\xc4\x42\x24\xc2\xa6\x8c\xdc\x1b\x76\x45\x18\x6b\xea\xe1\x94\x60\xd2\xa1\x5c\xcd\x61\xbd\x28\xe1\xd2\x01\x5f\x7d\x04\x5c\x38\x30\x8c\x5f\xf3\x0a\x66\xe6\xc8\xd2\x61\x0c\x22\x22\x52\x83\xd8\xdc\xa3\x8a\xf1\x16\xd2\x6e\xc1\x5e\x5a\x62\x01\x48\x41\xc7\xc6\xa5\xe5\xa6\x21\xb7\x8d\x66\x21\x50\x50\x60\x96\xfb\xc5\xaf\x29\xfb\x8b\xdf\xde\xf8\xa5\xf9\x76\x69\xe8\x3a\x30\x8d\xfa\xf0\x2d\x17\x3f\x55\x2e\xcb\xf6\x9d\xa5\x44\x02\xad\x2b\x1e\x13\xb1\x81\x47\xc1\x15\xe0\x13\xf3\x4e\x96\x7f\xa1\xf0\x9f\xde\xba\xc1\xed\x85\x14\xeb\x56\x65\x08\xfa\x7d\x6e\xc2\x90\x1e\x84\x39\xf9\xac\x21\xf4\x26\xb8\xfd\x7e\xf0\x1d\xa0\xe0\xa7\x89\xe5\xf8\x47\x56\xad\xd3\x8b\xf6\xa0\x09\x35\xa3\x7c\x22\xb7\x2e\x94\x3d\x70\xb7\x90\x1a\x08\xfd\x3c\xbf\xf9\xd9\x28\xbf\x86\xdb\x5a\xd9\xb4\x81\xa7\xe6\x43\xe5\xb8\xaa\xa5\x59\x13\x07\x45\xbd\x1b\xc7\x52\x8f\x81\x21\x21\x30\x45\xb2\xa9\x30\x31\xf9\xa1\xe0\xf3\x54\x2a\x88\xee\xa5\x24\x1b\x06\x61\xf7\xed\xf0\xac\x1c\xc4\x8a\xec\xe0\x1d\x98\xb7\xcc\x4c\xf2\x12\x64\x19\x77\xb2\x57\x5d\xee\x5d\xa2\x68\x2d\x69\x95\x5b\x6e\x21\xfd\x43\xd0\xc3\xb6\x6b\x10\xc0\x30\x8c\x7a\x05\x90\x0d\x67\x4b\x14\x2f\x40\xc9\xc4\x72\x8e\xe3\xa6\xcd\xde\x26\x94\x3a\x63\xef\xd8\x07\x2f\x8b\x99\xfc\x49\xd4\xf3\x61\x46\x0d\x39\xf7\x14\xaa\xa7\xc1\x1b\xff\xca\xd8\x4b\x96\xfc\xc8\xa3\x38\x51\xe0\x37\x3c\xbc\x0e\xf0\xce\x02\x77\x3a\xbb\x16\x78\xaf\x34\x06\x03\xb0\xdf\x97\xc3\x51\x68\x90\x1b\x3d\xc8\xb8\x26\xac\xfe\x15\xed\xf9\xda\xc4\xdc\x39\x8e\x68\x47\x42\x6d\x44\x5a\xc0\x96\xc3\x7f\x36\x75\x83\x37\x55\xf8\x17\xc4\xee\x9d\x88\xc6\xff\x42\x32\xf6\x14\x82\x9f\x04\x8a\xf5\x99\xd0\xc3\x34\x61\x85\xf1\xb5\x65\x63\x27\x25\x5e\xf7\xc4\x7f\xf1\x5c\xb4\xf3\xc1\x4d\xfe\xbc\xbf\xf9\x27\x00\x00\xff\xff\x87\xc4\xa9\xe3\xb1\x1b\x00\x00")
+var _clusterDevelopmentPredeployJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x58\xdf\x8f\xa3\x36\x10\x7e\xdf\xbf\x02\xd1\x4a\x24\x52\xf8\x91\xe8\x5a\x75\xf7\x6d\xd5\xf6\xaa\xd5\xa9\xd7\x55\x73\xdd\x3e\x44\x79\x70\xec\x49\xd6\x07\xd8\xc8\x36\xb9\x72\x55\xfe\xf7\xca\x10\x48\x00\x43\xb2\x34\xb9\x56\xab\xbb\x93\x56\xc2\x8c\xc7\x33\xdf\xcc\xf7\x79\xc2\xdf\x37\x96\x65\x59\xf6\xb7\x12\x3f\x43\x8c\xec\x3b\xcb\x7e\x56\x2a\x91\x77\xbe\x5f\xac\x78\x31\x62\x68\x03\x31\x30\xe5\xa1\xcf\xa9\x00\x0f\xf3\x78\xff\x4e\xfa\xb3\x60\xfa\x9d\x1b\x4c\xdd\x60\xea\x13\x48\x22\x9e\x69\xbb\x0f\x10\x27\x11\x52\xe0\x7d\x94\x9c\x7d\x63\x4f\x8a\x13\x30\x67\x0a\x98\x7a\x02\x21\x29\x67\xfa\xa0\xa9\x17\xe8\xff\xa5\x41\x82\x04\x8a\x41\x81\x90\xf6\x9d\x55\x84\x55\x6c\xa4\xb5\xe7\x7c\x4d\x65\x09\x68\x17\x2b\xce\xa3\xfd\xfe\xea\x1d\x81\x35\x4a\x23\xf5\x84\xa2\x54\xdb\xac\x51\x24\xa1\xb2\xd8\x4d\x8e\x1c\x47\xa9\x54\x20\xde\xa3\x18\xba\x4f\x90\x4a\x50\xb6\xb1\xfb\x1c\xcc\x41\x6c\x29\x86\x47\x41\x19\xa6\x09\x8a\x1e\xc8\x30\x77\xeb\xe4\x52\x9e\xc2\xed\xf0\xa4\x62\xa4\x73\xba\x27\x44\x80\x94\x8f\x02\xd6\xf4\xaf\x61\x8e\x04\x4f\x15\xc8\xee\xbd\x48\x08\x94\x9d\x28\xde\x62\x69\x74\xbd\x65\xa0\x2e\x10\xe1\x27\x2e\xc2\xe1\xa9\xde\x1c\xb9\xb3\x05\x48\x9e\x0a\x9c\xe7\xbb\xa8\x6c\x1a\xae\x12\xc1\x13\x10\x8a\x1a\x50\xc9\xdf\xa3\x22\x92\x79\x82\x70\xbb\x78\x4d\xab\x22\xde\xc6\x81\x2d\xe3\xc5\x81\x55\x23\xa7\x05\x9b\x33\x5e\xda\xc6\xbd\xcb\xd6\xea\xae\xb6\xb2\x6b\x94\x8d\x15\xed\x66\x13\xd8\xba\xfa\x94\x66\x59\x4b\x10\x7f\xa5\x58\x70\xc9\xd7\xca\x7b\x0f\x4a\xc3\xef\x6f\xa9\x50\x29\x8a\xf6\x8f\xb2\xb9\x31\xe2\x18\xa9\xbd\x5e\x2c\x4a\x90\x7f\x11\x3c\x4d\x46\x63\xaf\x7c\xb9\x6c\xee\xc2\x9c\x11\x5a\x6d\x3b\x86\x00\x53\x9d\x73\xc3\x1c\x25\xf4\x48\x96\x66\xc1\x2c\x70\x83\x1f\xdc\x60\x6a\xec\x9a\x17\xd6\xb4\x62\x41\x3d\x8e\x62\xb9\x85\x7f\x17\xae\x0b\xcc\x19\x46\x6a\x54\x4b\xe5\x20\x5f\xce\x78\x62\x39\xae\x50\x86\xdc\xba\x91\xcf\x43\xf8\x80\x56\x11\x5c\x08\xf5\x6b\xc2\x88\x1a\x24\xad\xa3\x69\xd0\xac\x36\x14\x87\x72\xe4\x49\x77\x13\x8c\x92\x5a\xde\x94\x8c\x9c\x5e\xf4\x9c\x89\x75\x6e\x79\x8c\x41\x15\x75\x42\x1b\x9d\x3a\x4b\xa3\x68\x20\xf7\xca\x1e\x71\x4a\x12\xfa\xce\xc4\xea\x8d\xa8\x80\xed\x45\x4d\xd3\xa0\xab\x2f\xd3\x15\x03\xd5\x6a\xa0\xce\x56\x68\xc9\x7d\x02\x8c\xc8\xdf\x98\x51\xc5\x4e\xd5\xa0\x11\x8c\x33\xb1\xaa\xdc\x3b\xca\x7f\xad\xa2\xd6\x4e\x5a\x0e\x63\xd3\xd5\x59\x62\xb8\xee\xbe\xb2\xe4\x34\x4b\x0a\xd8\x5e\x0d\x4b\xca\x60\x6a\x6c\x39\xaf\x36\xa5\x5e\xbc\x4a\x6e\x55\x0d\x72\x1c\x53\x31\x48\x9f\x55\xfb\x77\x90\x3d\xe9\xd1\xd5\xdf\xea\xbf\x17\xba\x52\x4f\x11\x5e\x01\x43\x4c\x3d\x14\x54\x94\xe9\x4a\x62\x41\x13\xed\x69\x34\xf6\xca\x77\xc6\x62\xc9\x30\xed\xa6\xf6\x1a\xc5\x34\xca\xb4\xcf\xfb\x2e\x22\x96\x68\x49\x85\x18\x41\x82\xb4\xc7\xc8\x9d\xe1\x54\x84\xb1\x96\x1e\x1e\x51\x5c\xa4\xb4\x58\x1a\xac\x80\xe9\x2e\x21\x6f\xb9\xf8\x89\xca\xf0\x67\x86\x45\x96\xec\xa1\x53\x22\x85\xce\x1d\x8f\xa9\xd8\xc0\xa3\xe0\x0a\xf0\x91\x79\xaf\x2a\xfc\xcb\x41\x71\x7a\xeb\x06\xb7\x17\x9a\x70\x42\x95\x23\xfe\xfb\xdc\x84\xb9\x1d\x42\x36\xa7\x9f\x35\xe4\x6f\x82\xdb\xef\x07\xcf\x8c\x65\x3f\x4f\x2c\xc7\x3f\xb0\xb0\xd9\x8e\xda\x83\x26\x20\xa1\x32\x74\xa1\xc2\xdf\x0d\x21\x33\x90\xff\x34\x17\xfc\x10\xb2\xff\xc7\x64\x5f\x15\x6c\xa0\xc2\x3e\xd4\xa4\xad\x91\x66\xe3\x22\x29\xb1\x6e\x49\xd8\x19\xcd\x42\x09\x30\x45\xf3\x8e\x30\xb1\x7e\x0f\xf8\x3c\x93\x0a\xe2\x7b\x29\xe9\x86\x01\xe9\xff\x25\x71\x72\x74\xc0\x8a\x6e\xe1\x1d\x98\x8f\xcc\x4d\x0a\x08\xf2\x8c\x3b\x8d\x2c\xc3\x68\x70\x09\xd0\x3a\xd2\xaa\x8e\x0c\x21\xfb\x43\x44\xfb\x63\xd7\x20\x80\x61\x18\x9d\x15\x40\xde\x9c\x1d\x51\x0c\x64\xc8\xc4\x72\x0e\xad\xa6\xef\xd9\xb7\x69\x14\x39\x63\xef\x50\x03\x2f\x8f\x97\xfe\x49\xd5\xf3\xbe\x3f\x0d\xf9\xbe\x6c\xa0\x39\x37\x44\x69\x1a\x8e\x0d\x1c\xfe\x91\xc7\x49\xaa\xc0\x27\x35\x0d\x9e\x1b\xc6\x98\xff\xea\x37\xfa\x74\x76\x2d\x26\x5f\xa1\x27\x06\x88\xc0\xb9\x42\x8e\x08\x19\x32\xa1\xf8\x8d\xdb\xf8\x8b\x0c\x2c\xed\x09\xc0\xc8\xe7\x1e\x71\x19\x3a\xf2\x54\x0e\xf8\xea\x23\xe0\xd2\xc1\x29\xb1\xe8\x66\xc1\xcb\xaf\xcf\x9c\x79\x85\x38\xcc\xa6\x6e\xf0\xa6\x2e\x0e\xa5\xe4\x7b\x47\x5f\x7f\x7b\xf3\x48\x40\xc4\x54\x6a\x46\x98\xb1\xae\x19\xe7\x37\x70\xdf\x97\xc2\xca\x72\xd3\xfa\x7a\x67\x34\xfb\x24\x50\xa2\x6f\x8b\x33\x4c\x53\x56\x1a\xf7\xda\xb6\xbf\x39\x96\xff\x76\xe6\xab\xa0\xb5\xba\xfc\x22\xb3\xde\x45\x04\xe7\xf2\x7d\xd5\x2d\x30\x37\xc5\xf3\xee\xe6\x9f\x00\x00\x00\xff\xff\x60\x27\x3a\x8d\xec\x19\x00\x00")
 
 func clusterDevelopmentPredeployJsonBytes() ([]byte, error) {
 	return bindataRead(
