+++ test/e2e/disk_encryption.go
@@ -16,8 +16,6 @@ import (
 )
 
 var _ = Describe("Encryption at host should be enabled", func() {
-	BeforeEach(skipIfNotInDevelopmentEnv)
-
 	Specify("each VM should have encryption at host enabled", func() {
 		ctx := context.Background()
 
@@ -33,7 +31,6 @@ var _ = Describe("Encryption at host should be enabled", func() {
 
 		By("checking the encryption property on each VM")
 		for _, vm := range vms {
-			log.Println(vm.Name)
 			Expect(vm.SecurityProfile).To(Not(BeNil()))
 			Expect(vm.SecurityProfile.EncryptionAtHost).To(Not(BeNil()))
 			Expect(*vm.SecurityProfile.EncryptionAtHost).To(Equal(true))
@@ -43,9 +40,6 @@ var _ = Describe("Encryption at host should be enabled", func() {
 })
 
 var _ = Describe("Disk encryption at rest should be enabled with customer managed key", func() {
-	BeforeEach(skipIfNotInDevelopmentEnv)
-
-	// We have to get the disks by VM, because when getting all disks by resource group, we do not get recently created disks, see https://github.com/Azure/azure-cli/issues/17123
 	Specify("each disk should have encryption at rest with customer managed key enabled", func() {
 		ctx := context.Background()
 
@@ -59,6 +53,8 @@ var _ = Describe("Disk encryption at rest should be enabled with customer manage
 		Expect(err).NotTo(HaveOccurred())
 		Expect(vms).NotTo(HaveLen(0))
 
+		// We have to get the disks by VM, because when getting all disks by resource group,
+		// we do not get recently created disks, see https://github.com/Azure/azure-cli/issues/17123
 		By("checking the encryption property on each OS disk of each VM")
 		for _, vm := range vms {
 			osDisk, err := clients.Disks.Get(ctx, clusterResourceGroup, *vm.StorageProfile.OsDisk.Name)
